Move django_amazon_ses to a single module instead of a deep package
The django_amazon_ses package was a couple directories deep but only
contain a single non-empty file. A single Python file in a package could
be more easily represented as a Python module. Makes configuration more
concise as well.
